renamed _registration
