+++ services/web/server/src/simcore_service_webserver/login/_registration.py
